feat: handle error responses to callback endpoint
This adds a check in middleware.ts to report an error if GitHub
sends the user back to the callback endpoint with an error query
parameter. Some potential error cases are redirect_uri_mismatch
for when a user visits the /oauth/authorize endpoint with a
redirect_uri query parameter that doesn't match the application's
and application_suspended for when the GitHub application is
suspended.