Don't respond to OPTIONS requests for non-oauth requests
This middleware is a little overzealous with its responses to `OPTIONS` requests. It really shouldn't respond to *every* request with these default highly permissive CORS headers.

Instead, it should only respond in this way if the path matches the prefix, otherwise defer to the application to respond to the `OPTIONS` request with its own CORS headers.